Issue #2325: Move PocketEndpoint.convertVideosJSON into PocketVideoParser; rm endpoint.
