+++ app/src/test/java/org/mozilla/tv/firefox/pocket/PocketVideoParserTest.kt
@@ -9,24 +9,17 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Assert.fail
-import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito.mock
 import org.mozilla.tv.firefox.TestResource
 import org.robolectric.RobolectricTestRunner
 
 private const val KEY_INNER = "recommendations"
 
 @RunWith(RobolectricTestRunner::class)
-class PocketEndpointTest {
+class PocketVideoParserTest {
 
-    private lateinit var pocketEndpoint: PocketEndpoint
-
-    @Before
-    fun setup() {
-        pocketEndpoint = PocketEndpoint(mock(PocketEndpointRaw::class.java)) { true }
-    }
+    private val pocketVideoParser = PocketVideoParser
 
     @Test
     fun `convert Videos JSON successfully to internal objects`() {
@@ -51,7 +44,7 @@ class PocketEndpointTest {
         )
 
         val pocketJSON = TestResource.POCKET_VIDEO_RECOMMENDATION.get()
-        val actualVideos = pocketEndpoint.convertVideosJSON(pocketJSON)
+        val actualVideos = pocketVideoParser.convertVideosJSON(pocketJSON)
         if (actualVideos == null) { fail("Expected actualVideos to be non-null"); return }
 
         // We only test a subset of the data for developer sanity. :)
@@ -69,7 +62,7 @@ class PocketEndpointTest {
         assertNotNull(expectedFirstTitle)
 
         val pocketJSONWithNoTitleExceptFirst = removeTitleStartingAtIndex(1, pocketJSON)
-        val actualVideos = pocketEndpoint.convertVideosJSON(pocketJSONWithNoTitleExceptFirst)
+        val actualVideos = pocketVideoParser.convertVideosJSON(pocketJSONWithNoTitleExceptFirst)
         if (actualVideos == null) { fail("Expected videos non-null"); return }
         assertEquals(1, actualVideos.size)
         assertEquals(expectedFirstTitle, actualVideos[0].title)
@@ -79,18 +72,18 @@ class PocketEndpointTest {
     fun `convert Videos JSON for videos with missing fields on all items`() {
         val pocketJSON = TestResource.POCKET_VIDEO_RECOMMENDATION.get()
         val pocketJSONWithNoTitles = removeTitleStartingAtIndex(0, pocketJSON)
-        val actualVideos = pocketEndpoint.convertVideosJSON(pocketJSONWithNoTitles)
+        val actualVideos = pocketVideoParser.convertVideosJSON(pocketJSONWithNoTitles)
         assertNull(actualVideos)
     }
 
     @Test
     fun `convert Videos JSON for empty String`() {
-        assertNull(pocketEndpoint.convertVideosJSON(""))
+        assertNull(pocketVideoParser.convertVideosJSON(""))
     }
 
     @Test
     fun `convert Videos JSON for invalid JSON`() {
-        assertNull(pocketEndpoint.convertVideosJSON("{!!}}"))
+        assertNull(pocketVideoParser.convertVideosJSON("{!!}}"))
     }
 }
